Add VerifyError and dedicated Error/ErrorKind types
The reasoning behind these changes is to be able to preserve additional
information behind the underlying `verify::ErrorCode`, specifically
whether it is a code mismatch error or not. This is a stepping stone for
statically guaranteeing `PendingVerify::check` retries.

In the process of making these changes, we remove the `AuthError` type.